Removed old tests, moved things to where they belong and fixed travis
