+++ gfe_integration_tests/test_rte/src/test_rte_start.c
