+++ gfe_integration_tests/test_rte/test_rte_during_run.mk
@@ -1,8 +1,12 @@
+# If SPINN_DIRS is not defined, this is an error!
+ifndef SPINN_DIRS
+    $(error SPINN_DIRS is not set.  Please define SPINN_DIRS (possibly by running "source setup" in the spinnaker package folder))
+endif
+
 APP = test_rte_during_run
-APP_OUTPUT_DIR = ./
-BUILD_DIR = build/
-SOURCE_DIRS += $(abspath .)
 SOURCES = test_rte_during_run.c
 
+APP_OUTPUT_DIR := $(abspath $(dir $(abspath $(lastword $(MAKEFILE_LIST)))))/
+
 # The spinnaker_tools standard makefile
-include $(SPINN_DIRS)/make/Makefile.SpiNNFrontEndCommon
+include $(SPINN_DIRS)/make/local.mk
