+++ spinnaker_graph_front_end/examples/template/src/c_template_vertex.c
