--- spinnaker_graph_front_end/examples/test_fixed_router_transmitter_reciever/Makefile
@@ -1,8 +1,0 @@
-all: hello_world.aplx hello_world_clone.aplx
-
-%.aplx: %.Makefile src/%.c
-	"$(MAKE)" -f $<
-
-clean:
-	for d in hello_world hello_world_clone ; do ("$(MAKE)" -f $$d.Makefile clean) || exit $$?; done
-
