--- spinnaker_graph_front_end/examples/test_fixed_router_transmitter_reciever/hello_world_clone.Makefile
@@ -1,11 +1,0 @@
-# If SPINN_DIRS is not defined, this is an error!
-ifndef SPINN_DIRS
-    $(error SPINN_DIRS is not set.  Please define SPINN_DIRS (possibly by running "source setup" in the spinnaker package folder))
-endif
-
-APP = hello_world_clone
-SOURCES = hello_world_clone.c
-
-APP_OUTPUT_DIR := $(abspath $(dir $(abspath $(lastword $(MAKEFILE_LIST)))))/
-
-include $(SPINN_DIRS)/make/local.mk
